+++ packages/to-json-schema/src/converters/convertAction/convertAction.ts
@@ -1,7 +1,7 @@
 import type { JSONSchema7 } from 'json-schema';
 import type * as v from 'valibot';
-import type { ConversionConfig } from './type.ts';
-import { handleError } from './utils/index.ts';
+import type { ConversionConfig } from '../../type.ts';
+import { addError, handleError } from '../../utils/index.ts';
 
 /**
  * Action type.
@@ -65,6 +65,7 @@ type Action =
       v.ValueInput,
       v.ErrorMessage<v.MaxValueIssue<v.ValueInput, v.ValueInput>> | undefined
     >
+  | v.MetadataAction<unknown, Record<string, unknown>>
   | v.MinEntriesAction<
       v.EntriesInput,
       number,
@@ -116,6 +117,15 @@ export function convertAction(
   valibotAction: Action,
   config: ConversionConfig | undefined
 ): JSONSchema7 {
+  // Ignore action if specified in configuration
+  if (config?.ignoreActions?.includes(valibotAction.type)) {
+    return jsonSchema;
+  }
+
+  // Create errors variable
+  let errors: [string, ...string[]] | undefined;
+
+  // Convert Valibot action to JSON Schema
   switch (valibotAction.type) {
     case 'base64': {
       jsonSchema.contentEncoding = 'base64';
@@ -151,9 +161,9 @@ export function convertAction(
         jsonSchema.maxItems = 0;
       } else {
         if (jsonSchema.type !== 'string') {
-          handleError(
-            `The "${valibotAction.type}" action is not supported on type "${jsonSchema.type}".`,
-            config
+          errors = addError(
+            errors,
+            `The "${valibotAction.type}" action is not supported on type "${jsonSchema.type}".`
           );
         }
         jsonSchema.maxLength = 0;
@@ -204,9 +214,9 @@ export function convertAction(
         jsonSchema.maxItems = valibotAction.requirement;
       } else {
         if (jsonSchema.type !== 'string') {
-          handleError(
-            `The "${valibotAction.type}" action is not supported on type "${jsonSchema.type}".`,
-            config
+          errors = addError(
+            errors,
+            `The "${valibotAction.type}" action is not supported on type "${jsonSchema.type}".`
           );
         }
         jsonSchema.minLength = valibotAction.requirement;
@@ -225,9 +235,9 @@ export function convertAction(
         jsonSchema.maxItems = valibotAction.requirement;
       } else {
         if (jsonSchema.type !== 'string') {
-          handleError(
-            `The "${valibotAction.type}" action is not supported on type "${jsonSchema.type}".`,
-            config
+          errors = addError(
+            errors,
+            `The "${valibotAction.type}" action is not supported on type "${jsonSchema.type}".`
           );
         }
         jsonSchema.maxLength = valibotAction.requirement;
@@ -237,13 +247,26 @@ export function convertAction(
 
     case 'max_value': {
       if (jsonSchema.type !== 'number') {
-        handleError(
-          `The "max_value" action is not supported on type "${jsonSchema.type}".`,
-          config
+        errors = addError(
+          errors,
+          `The "max_value" action is not supported on type "${jsonSchema.type}".`
         );
       }
       // @ts-expect-error
       jsonSchema.maximum = valibotAction.requirement;
+      break;
+    }
+
+    case 'metadata': {
+      if (typeof valibotAction.metadata.title === 'string') {
+        jsonSchema.title = valibotAction.metadata.title;
+      }
+      if (typeof valibotAction.metadata.description === 'string') {
+        jsonSchema.description = valibotAction.metadata.description;
+      }
+      if (Array.isArray(valibotAction.metadata.examples)) {
+        jsonSchema.examples = valibotAction.metadata.examples;
+      }
       break;
     }
 
@@ -257,9 +280,9 @@ export function convertAction(
         jsonSchema.minItems = valibotAction.requirement;
       } else {
         if (jsonSchema.type !== 'string') {
-          handleError(
-            `The "${valibotAction.type}" action is not supported on type "${jsonSchema.type}".`,
-            config
+          errors = addError(
+            errors,
+            `The "${valibotAction.type}" action is not supported on type "${jsonSchema.type}".`
           );
         }
         jsonSchema.minLength = valibotAction.requirement;
@@ -269,9 +292,9 @@ export function convertAction(
 
     case 'min_value': {
       if (jsonSchema.type !== 'number') {
-        handleError(
-          `The "min_value" action is not supported on type "${jsonSchema.type}".`,
-          config
+        errors = addError(
+          errors,
+          `The "min_value" action is not supported on type "${jsonSchema.type}".`
         );
       }
       // @ts-expect-error
@@ -289,9 +312,9 @@ export function convertAction(
         jsonSchema.minItems = 1;
       } else {
         if (jsonSchema.type !== 'string') {
-          handleError(
-            `The "${valibotAction.type}" action is not supported on type "${jsonSchema.type}".`,
-            config
+          errors = addError(
+            errors,
+            `The "${valibotAction.type}" action is not supported on type "${jsonSchema.type}".`
           );
         }
         jsonSchema.minLength = 1;
@@ -301,7 +324,10 @@ export function convertAction(
 
     case 'regex': {
       if (valibotAction.requirement.flags) {
-        handleError('RegExp flags are not supported by JSON Schema.', config);
+        errors = addError(
+          errors,
+          'RegExp flags are not supported by JSON Schema.'
+        );
       }
       jsonSchema.pattern = valibotAction.requirement.source;
       break;
@@ -332,12 +358,33 @@ export function convertAction(
     }
 
     default: {
-      handleError(
+      errors = addError(
+        errors,
         // @ts-expect-error
-        `The "${valibotAction.type}" action cannot be converted to JSON Schema.`,
-        config
+        `The "${valibotAction.type}" action cannot be converted to JSON Schema.`
       );
     }
   }
+
+  // Override JSON Schema if specified and necessary
+  if (config?.overrideAction) {
+    const actionOverride = config.overrideAction({
+      valibotAction,
+      jsonSchema,
+      errors,
+    });
+    if (actionOverride) {
+      return { ...actionOverride };
+    }
+  }
+
+  // Handle errors based on configuration
+  if (errors) {
+    for (const message of errors) {
+      handleError(message, config);
+    }
+  }
+
+  // Return converted JSON Schema
   return jsonSchema;
 }