+++ packages/to-json-schema/src/functions/index.ts
@@ -1,0 +1,2 @@
+export * from './toJsonSchema/index.ts';
+export * from './toJsonSchemaDefs/index.ts';